requirements.txt
streamlit>=1.33.0
google-generativeai>=0.7.2
requests>=2.31.0
beautifulsoup4>=4.12.0
exa-py>=1.0.0
python-dotenv>=1.0.0


